Added a comprehensive details page (SEE DESCRIPTION!!!) Added favorites page.
On the details page, I thought being able to use css to tab within a div
would be awesome. It worked out well (I think) save for one thing: The
height of the div is fixed, and I have not found a way to make it
responsive yet. This could be worth revisiting in the future.