Changed quite a bit. As much polishing as I can get done for now. (See Description)
So this is about as far as I can get with the skills I have. There are
limited validation errors across a couple of the pages, and these are
things that I'm either not sure how to fix while keeping the effect that
I feel is necessary to the page, or I don't know how to fix it at all.